delete instrumentation for EsRestClient v7.1.0, added AbstractEsClientInstrumentationTest.
