+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/test/java/co/elastic/apm/agent/es/restclient/v6_4/ElasticsearchRestClientInstrumentationIT.java
@@ -24,13 +24,8 @@
  */
 package co.elastic.apm.agent.es.restclient.v6_4;
 
-import co.elastic.apm.agent.AbstractInstrumentationTest;
-import co.elastic.apm.agent.impl.error.ErrorCapture;
-import co.elastic.apm.agent.impl.transaction.Db;
-import co.elastic.apm.agent.impl.transaction.Http;
+import co.elastic.apm.agent.es.restclient.AbstractEsClientInstrumentationTest;
 import co.elastic.apm.agent.impl.transaction.Span;
-import co.elastic.apm.agent.impl.transaction.TraceContext;
-import co.elastic.apm.agent.impl.transaction.Transaction;
 import org.apache.http.HttpHost;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
@@ -59,9 +54,7 @@ import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -69,47 +62,23 @@ import org.junit.runners.Parameterized;
 import org.testcontainers.elasticsearch.ElasticsearchContainer;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.ELASTICSEARCH;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SEARCH_QUERY_PATH_SUFFIX;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_ACTION;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
 
 @RunWith(Parameterized.class)
-public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrumentationTest {
-    private static final String USER_NAME = "elastic-user";
-    private static final String PASSWORD = "elastic-pass";
+public class ElasticsearchRestClientInstrumentationIT extends AbstractEsClientInstrumentationTest {
 
-    @SuppressWarnings("NullableProblems")
-    private static ElasticsearchContainer container;
     @SuppressWarnings("NullableProblems")
     private static RestHighLevelClient client;
 
-    private static final String INDEX = "my-index";
-    private static final String SECOND_INDEX = "my-second-index";
-    private static final String DOC_ID = "38uhjds8s4g";
-    private static final String DOC_TYPE = "_doc";
-    private static final String FOO = "foo";
-    private static final String BAR = "bar";
-    private static final String BAZ = "baz";
-
-    private boolean async;
-
     public ElasticsearchRestClientInstrumentationIT(boolean async) {
         this.async = async;
-    }
-
-    @Parameterized.Parameters(name = "Async={0}")
-    public static Iterable<Object[]> data() {
-        return Arrays.asList(new Object[][]{{Boolean.FALSE}, {Boolean.TRUE}});
     }
 
     @BeforeClass
@@ -137,23 +106,6 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
         client.close();
     }
 
-    @Before
-    public void startTransaction() {
-        Transaction transaction = tracer.startTransaction(TraceContext.asRoot(), null, null).activate();
-        transaction.setName("ES Transaction");
-        transaction.withType("request");
-        transaction.withResult("success");
-    }
-
-    @After
-    public void endTransaction() {
-        Transaction currentTransaction = tracer.currentTransaction();
-        if (currentTransaction != null) {
-            currentTransaction.deactivate().end();
-        }
-        reporter.reset();
-    }
-
     @Test
     public void testTryToDeleteNonExistingIndex() throws IOException, InterruptedException {
         ElasticsearchStatusException ese = null;
@@ -169,48 +121,7 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
         assertThat(ese).isNotNull();
         assertThat(ese.status().getStatus()).isEqualTo(404);
 
-        System.out.println(reporter.generateErrorPayloadJson());
-
-        List<ErrorCapture> errorCaptures = reporter.getErrors();
-        assertThat(errorCaptures).hasSize(1);
-        ErrorCapture errorCapture = errorCaptures.get(0);
-        assertThat(errorCapture.getException()).isNotNull();
-/*
-        Map<String, String> tags = errorCapture.getContext().getLabels();
-        assertThat(tags).containsKey(QUERY_STATUS_CODE_KEY);
-        assertThat(tags.get(QUERY_STATUS_CODE_KEY)).isEqualTo("404");
-        assertThat(tags).containsKey(ERROR_REASON_KEY);
-        assertThat(tags).containsKey(ELASTICSEARCH_NODE_URL_KEY);
-        assertThat(tags.get(ELASTICSEARCH_NODE_URL_KEY)).isEqualTo(container.getHost().toURI().toString());
-*/
-    }
-
-    private void validateSpanContent(Span span, String expectedName, int statusCode, String method) {
-        assertThat(span.getType()).isEqualTo(SPAN_TYPE);
-        assertThat(span.getSubtype()).isEqualTo(ELASTICSEARCH);
-        assertThat(span.getAction()).isEqualTo(SPAN_ACTION);
-        assertThat(span.getName().toString()).isEqualTo(expectedName);
-        validateHttpContextContent(span.getContext().getHttp(), statusCode, method);
-
-        assertThat(span.getContext().getDb().getType()).isEqualTo(ELASTICSEARCH);
-
-        if (!expectedName.contains(SEARCH_QUERY_PATH_SUFFIX)) {
-            assertThat((CharSequence) (span.getContext().getDb().getStatementBuffer())).isNull();
-        }
-    }
-
-    private void validateHttpContextContent(Http http, int statusCode, String method) {
-        assertThat(http).isNotNull();
-        assertThat(http.getMethod()).isEqualTo(method);
-        assertThat(http.getStatusCode()).isEqualTo(statusCode);
-        assertThat(http.getUrl()).isEqualTo("http://" + container.getHttpHostAddress());
-    }
-
-    private void validateDbContextContent(Span span, String statement) {
-        Db db = span.getContext().getDb();
-        assertThat(db.getType()).isEqualTo(ELASTICSEARCH);
-        assertThat((CharSequence) db.getStatementBuffer()).isNotNull();
-        assertThat(db.getStatementBuffer().toString()).isEqualTo(statement);
+        assertThatErrorsExistWhenDeleteNonExistingIndex();
     }
 
     @Test
@@ -218,22 +129,13 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
         // Create an Index
         doCreateIndex(new CreateIndexRequest(SECOND_INDEX));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        List<Span> spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: PUT /%s", SECOND_INDEX), 200, "PUT");
-
+        validateSpanContentAfterIndexCreateRequest();
         // Delete the index
         reporter.reset();
 
         doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: DELETE /%s", SECOND_INDEX), 200, "DELETE");
+        validateSpanContentAfterIndexDeleteRequest();
     }
 
     @Test
@@ -318,11 +220,7 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
             ))
             .add(new DeleteRequest(INDEX, DOC_TYPE, "2")));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        List<Span> spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        assertThat(spans.get(0).getName().toString()).isEqualTo("Elasticsearch: POST /_bulk");
+        validateSpanContentAfterBulkRequest();
     }
 
     private interface ClientMethod<Req, Res> {
