--- apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-7_1/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,2 +1,0 @@
-co.elastic.apm.agent.es.restclient.v7_1.ElasticsearchClientSyncInstrumentation
-co.elastic.apm.agent.es.restclient.v7_1.ElasticsearchClientAsyncInstrumentation
