+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-7_1/src/test/java/co/elastic/apm/agent/es/restclient/v7_1/ElasticsearchRestClientInstrumentationIT.java
@@ -1,14 +1,10 @@
 package co.elastic.apm.agent.es.restclient.v7_1;
 
-import co.elastic.apm.agent.AbstractInstrumentationTest;
-import co.elastic.apm.agent.impl.error.ErrorCapture;
-import co.elastic.apm.agent.impl.transaction.Db;
-import co.elastic.apm.agent.impl.transaction.Http;
+import co.elastic.apm.agent.es.restclient.AbstractEsClientInstrumentationTest;
 import co.elastic.apm.agent.impl.transaction.Span;
-import co.elastic.apm.agent.impl.transaction.TraceContext;
-import co.elastic.apm.agent.impl.transaction.Transaction;
 import org.apache.http.HttpHost;
 import org.apache.http.impl.client.BasicCredentialsProvider;
+import org.elasticsearch.ElasticsearchStatusException;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.action.bulk.BulkRequest;
@@ -25,9 +21,6 @@ import org.elasticsearch.action.update.UpdateRequest;
 import org.elasticsearch.action.update.UpdateResponse;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.RequestOptions;
-import org.elasticsearch.client.Response;
-import org.elasticsearch.client.ResponseException;
-import org.elasticsearch.client.ResponseListener;
 import org.elasticsearch.client.RestClient;
 import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
@@ -35,9 +28,7 @@ import org.elasticsearch.client.indices.CreateIndexRequest;
 import org.elasticsearch.client.indices.CreateIndexResponse;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -48,50 +39,21 @@ import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.ELASTICSEARCH;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SEARCH_QUERY_PATH_SUFFIX;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_ACTION;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
 
 @RunWith(Parameterized.class)
-public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrumentationTest {
-
-    private static final String USER_NAME = "elastic-user";
-    private static final String PASSWORD = "elastic-pass";
-
-    @SuppressWarnings("NullableProblems")
-    private static ElasticsearchContainer container;
-
-    @SuppressWarnings("NullableProblems")
-    private static RestClient lowLevelClient;
+public class ElasticsearchRestClientInstrumentationIT extends AbstractEsClientInstrumentationTest {
 
     private static RestHighLevelClient client;
 
-    private static final String INDEX = "my-index";
-    private static final String SECOND_INDEX = "my-second-index";
-    private static final String DOC_TYPE = "_doc";
-    private static final String DOC_ID = "38uhjds8s5g";
-    private static final String FOO = "foo";
-    private static final String BAR = "bar";
-    private static final String BAZ = "baz";
-
-    private boolean async;
-
     public ElasticsearchRestClientInstrumentationIT(boolean async) { this.async = async; }
-
-    @Parameterized.Parameters(name = "Async={0}")
-    public static Iterable<Object[]> data() {
-        return Arrays.asList(new Object[][]{{Boolean.FALSE}, {Boolean.TRUE}});
-    }
 
     @BeforeClass
     public static void startElasticsearchContainerAndClient() throws IOException {
@@ -120,119 +82,38 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
         lowLevelClient.close();;
     }
 
-    @Before
-    public void startTransaction() {
-        Transaction transaction = tracer.startTransaction(TraceContext.asRoot(), null, null).activate();
-        transaction.setName("ES Transaction");
-        transaction.withType("request");
-        transaction.withResult("success");
-    }
-
-    @After
-    public void endTransaction() {
-        try {
-            Transaction currentTransaction = tracer.currentTransaction();
-            if (currentTransaction != null) {
-                currentTransaction.deactivate().end();
-            }
-        } finally {
-            reporter.reset();
-        }
-    }
-
-    private Response doPerformRequest(String method, String path) throws IOException, ExecutionException {
-        Request request = new Request(method, path);
-        if (async) {
-            final CompletableFuture<Response> resultFuture = new CompletableFuture<>();
-
-            lowLevelClient.performRequestAsync(request, new ResponseListener() {
-                @Override
-                public void onSuccess(Response response) {
-                    resultFuture.complete(response);
-                }
-
-                @Override
-                public void onFailure(Exception e) {
-                    resultFuture.completeExceptionally(e);
-                }
-            });
-            try {
-                return resultFuture.get();
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-        return lowLevelClient.performRequest(request);
-    }
-
     @Test
     public void testTryToDeleteNonExistingIndex() throws IOException {
-        ResponseException re = null;
+        ElasticsearchStatusException ese = null;
         try {
-            doPerformRequest("POST", "/non-existing/1/_mapping");
-        } catch (ResponseException e) {
-            re = e;
+            doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
+        } catch (ElasticsearchStatusException e) {
+            // sync scenario
+            ese = e;
         } catch (ExecutionException e) {
-            re = (ResponseException) e.getCause();
+            // async scenario
+            ese = (ElasticsearchStatusException) e.getCause();
+        } catch (InterruptedException e) {
+            e.printStackTrace();
         }
-        assertThat(re).isNotNull();
-        assertThat(re.getResponse().getStatusLine().getStatusCode()).isEqualTo(400);
+        assertThat(ese).isNotNull();
+        assertThat(ese.status().getStatus()).isEqualTo(404);
 
-        System.out.println(reporter.generateErrorPayloadJson());
-
-        List<ErrorCapture> errorCaptures = reporter.getErrors();
-        assertThat(errorCaptures).hasSize(1);
-        ErrorCapture errorCapture = errorCaptures.get(0);
-        assertThat(errorCapture.getException()).isNotNull();
+        assertThatErrorsExistWhenDeleteNonExistingIndex();
     }
 
-    private void validateSpanContent(Span span, String expectedName, int statusCode, String method) {
-        assertThat(span.getType()).isEqualTo(SPAN_TYPE);
-        assertThat(span.getSubtype()).isEqualTo(ELASTICSEARCH);
-        assertThat(span.getAction()).isEqualTo(SPAN_ACTION);
-        assertThat(span.getName().toString()).isEqualTo(expectedName);
-        validateHttpContextContent(span.getContext().getHttp(), statusCode, method);
-
-        assertThat(span.getContext().getDb().getType()).isEqualTo(ELASTICSEARCH);
-
-        if (!expectedName.contains(SEARCH_QUERY_PATH_SUFFIX)) {
-            assertThat((CharSequence) (span.getContext().getDb().getStatementBuffer())).isNull();
-        }
-    }
-
-    private void validateHttpContextContent(Http http, int statusCode, String method) {
-        assertThat(http).isNotNull();
-        assertThat(http.getMethod()).isEqualTo(method);
-        assertThat(http.getStatusCode()).isEqualTo(statusCode);
-        assertThat(http.getUrl()).isEqualTo("http://" + container.getHttpHostAddress());
-    }
-
-    private void validateDbContextContent(Span span, String statement) {
-        Db db = span.getContext().getDb();
-        assertThat(db.getType()).isEqualTo(ELASTICSEARCH);
-        assertThat((CharSequence) db.getStatementBuffer()).isNotNull();
-        assertThat(db.getStatementBuffer().toString()).isEqualTo(statement);
-    }
 
     @Test
     public void testCreateAndDeleteIndex() throws InterruptedException, ExecutionException, IOException {
         doCreateIndex(new CreateIndexRequest(SECOND_INDEX));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        List<Span> spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: PUT /%s", SECOND_INDEX), 200, "PUT");
+        validateSpanContentAfterIndexCreateRequest();
 
         reporter.reset();
 
         doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: DELETE /%s", SECOND_INDEX), 200, "DELETE");
+        validateSpanContentAfterIndexDeleteRequest();
     }
 
     @Test
@@ -315,11 +196,7 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
             ))
             .add(new DeleteRequest(INDEX, DOC_TYPE, "2")));
 
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        List<Span> spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        assertThat(spans.get(0).getName().toString()).isEqualTo("Elasticsearch: POST /_bulk");
+        validateSpanContentAfterBulkRequest();
     }
 
     private interface ClientMethod<Req, Res> {
@@ -404,6 +281,4 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractInstrument
         }
         return client.bulk(bulkRequest, RequestOptions.DEFAULT);
     }
-
-
 }
